- composite grading scale with grades
